+++ src/hooks/User.tsx
@@ -24,10 +24,15 @@ export const useUserBadges = () =>
     getNextPageParam: (lastPage) => lastPage.next,
   });
 
-export const useUserEvents = () =>
-  useInfiniteQuery<PaginationResponse<EventCompact>, RequestResponse>([USER_EVENTS_QUERY_KEY], ({ pageParam = 1 }) => API.getUserEvents({ page: pageParam }), {
-    getNextPageParam: (lastPage) => lastPage.next,
-  });
+export const useUserEvents = () => {
+  return useInfiniteQuery<PaginationResponse<EventCompact>, RequestResponse>(
+    [USER_EVENTS_QUERY_KEY],
+    ({ pageParam = 1 }) => API.getUserEvents({ page: pageParam }),
+    {
+      getNextPageParam: (lastPage) => lastPage.next,
+    },
+  );
+};
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const useUserForms = (filters?: any) =>